Add note about where the valueFiles param references (#3196)
This file seems to serve as the only real reference as to what parameters are available for the Application CRD. It took me a bit of trial and error to figure out that these are in the git repo referenced and relative to the path of the helm chart set previously. Hopefully this helps a future user save some time.