fix(ui): source can in fact be `undefined` (#20381)
* fix(ui): source can in fact be `undefined`

The assumption that a source is always there is not always true. To
repro, create an app-of-apps containing a single app without any `source`
present. In the UI this will crash, horribly. This PR fixes that so
that instead of crashing the user will get useful info indicating what
is wrong with the app.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>

* chore(ui): some cr tweaks

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>

* chore(ui): some cr tweaks

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>

---------

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>